+++ package.json
@@ -98,6 +98,7 @@
     "i18next-browser-languagedetector": "^6.1.3",
     "i18next-http-backend": "^1.3.2",
     "react": "^17.0.2",
+    "react-device-detect": "^2.2.3",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
     "react-i18next": "^11.15.5",
