+++ src/entities/Lesta/index.ts
@@ -16,7 +16,7 @@ export {
   LestaUserSession,
 } from './model/types/users';
 export { LestaClan } from './model/types/clans';
-export { LestaTankStats, LestaTankData } from './model/types/tanks';
+export { LestaTankStats, LestaTankData, TUserTanks } from './model/types/tanks';
 export { fetchLestaUserDataById } from './model/services/fetchLestaUserDataById/fetchLestaUserDataById';
 export { createLestaUserSession } from './model/services/createLestaUserSession/createLestaUserSession';
 export { ParamData } from './model/types/default';
