+++ src/entities/Tank/index.ts
@@ -1,1 +1,7 @@
 export { Tank } from './ui/Tank/Tank';
+export { getLevelRoman } from './lib/converterTank';
+export {
+  masteryTank,
+  nationFlag,
+  typeIcon,
+} from './config/TankData';
