+++ src/features/Filter/lib/getFilterData/getFilterItem.ts
@@ -1,0 +1,5 @@
+import { filterData } from '../../config/filterData';
+
+export const getFilterItem = (tab: number) => (tab === 1
+  ? filterData.filter((item) => item.nameParam !== 'mark_of_mastery')
+  : filterData);
