--- src/features/Filter/model/selectors/getCheckboxesFilterState/getCheckboxesFilterState.ts
@@ -1,3 +1,0 @@
-import { StateSchema } from 'app/providers/StoreProvider';
-
-export const getCheckboxesFilterState = (state: StateSchema) => state?.filter?.checkboxes || {};
