--- src/features/Filter/model/selectors/getDataFilterState/getDataFilterState.ts
@@ -1,3 +1,0 @@
-import { StateSchema } from 'app/providers/StoreProvider';
-
-export const getDataFilterState = (state: StateSchema) => state?.filter?.data || [];
