--- src/features/Filter/model/selectors/getIsActiveFilter/getIsActiveFilter.ts
@@ -1,3 +1,0 @@
-import { StateSchema } from 'app/providers/StoreProvider/config/StateSchema';
-
-export const getIsActiveFilter = (state: StateSchema) => state?.filter?.isActiveFilter || false;
