--- src/features/Filter/model/selectors/getSortList/getSortList.ts
@@ -1,3 +1,0 @@
-import { StateSchema } from 'app/providers/StoreProvider';
-
-export const getSortList = (state: StateSchema) => state?.sort;
