+++ src/features/Filter/model/selectors/index.ts
@@ -1,0 +1,6 @@
+import { StateSchema } from 'app/providers/StoreProvider';
+
+export const getCheckboxesFilterState = (state: StateSchema) => state?.filter?.checkboxes || {};
+export const getDataFilterState = (state: StateSchema) => state?.filter?.data || [];
+export const getIsActiveFilter = (state: StateSchema) => state?.filter?.isActiveFilter || false;
+export const getSortList = (state: StateSchema) => state?.sort;
